+++ src/modules/main/store/actionTypes.ts
