+++ src/store/rootReducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from 'redux';
 
-import { directionReducer } from '../modules/direction/directionReducer/directionReducer';
-import { mainReducer } from '../modules/main/mainReducer/mainReducer';
+import { directionReducer } from '../modules/direction/store/directionReducer';
+import { mainReducer } from '../modules/main/store/mainReducer';
 
 export default combineReducers({
   main: mainReducer,
